Add tests for adding
