+++ package.json
@@ -4,6 +4,7 @@
   "description": "Simple todo list app",
   "private": true,
   "scripts": {
+    "test": "jest",
     "deploy": "gh-pages -d dist",
     "build": "webpack",
     "start": "webpack serve --open"
@@ -20,6 +21,7 @@
   },
   "homepage": "https://github.com/IndieCoderMM/todo-app",
   "devDependencies": {
+    "@babel/plugin-transform-modules-commonjs": "^7.19.6",
     "babel-eslint": "^10.1.0",
     "css-loader": "^6.7.2",
     "eslint": "^7.32.0",
@@ -27,6 +29,8 @@
     "eslint-plugin-import": "^2.26.0",
     "hint": "^7.1.3",
     "html-webpack-plugin": "^5.5.0",
+    "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
     "style-loader": "^3.3.1",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
